[Add]カラム修正

+++ db/migrate/20210911074111_create_conditions.rb
@@ -2,7 +2,7 @@ class CreateConditions < ActiveRecord::Migration[5.2]
   def change
     create_table :conditions do |t|
       t.integer :user_id
-      t.integer :temperature
+      t.float :temperature
       t.boolean :condition, default: true, null: false
       t.string :remark
